pyproject.toml
[tool.poetry]
name = "home-price-prediction"
version = "0.1.0"
description = ""
authors = ["IAmLamarr <vitdub2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
dvc-gdrive = "^3.0.0"
matplotlib = "^3.8.2"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
skl2onnx = "^1.16.0"
onnxruntime = "^1.16.3"
pre-commit = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "hug"

[tool.flake8]
profile = "hug"
